+++ library/Perfdatagraphs/Common/PerfdataChart.php
@@ -4,14 +4,18 @@ namespace Icinga\Module\Perfdatagraphs\Common;
 
 use Icinga\Module\Perfdatagraphs\Widget\QuickActions;
 
-use ipl\Html\HtmlElement;
+use Icinga\Util\Json;
+
 use ipl\Html\Html;
+use ipl\Html\HtmlElement;
 use ipl\Html\ValidHtml;
-use ipl\Web\Widget\Icon;
 use ipl\I18n\Translation;
+use ipl\Web\Url;
+use ipl\Web\Widget\Icon;
 
 /**
  * PerfdataChart contains common functionality used for rendering the performance data charts.
+ * The idea is that you use this in the hook to create the chart elements.
  */
 trait PerfdataChart
 {
@@ -72,6 +76,7 @@ trait PerfdataChart
         // Where we store all elements for the charts.
         $charts = HtmlElement::create('div', [
             'class' => 'perfdata-charts-container collapsible',
+            'id' => $elemID,
             // Note: We could have a configuration option to change the
             // "always collapsed" behaviour
             'data-visible-height' => 0,
@@ -85,47 +90,104 @@ trait PerfdataChart
             'id' => $elemID . '-control',
         ]);
 
-        $b = new HtmlElement(
+        $toggleButton = new HtmlElement(
             'button',
             null,
             new Icon('angle-double-up', ['class' => 'collapse-icon']),
             new Icon('angle-double-down', ['class' => 'expand-icon'])
         );
 
-        $chartsControl->add($b);
-
-        // Element in which the charts will get rendered.
-        // We use attributes on this elements to transport data
-        // to the JavaScript part of this module.
-        $chart = HtmlElement::create('div', [
-            'id' => $elemID,
-            'class' => 'line-chart',
-            'data-host' => $hostName,
-            'data-ishostcheck' => $isHostCheck ? 'true': 'false',
-            'data-service' => $serviceName,
-            'data-checkcommand' => $checkCommandName,
-        ]);
-
-        // This element can be used to show error messages when fetching data fails.
-        $error = HtmlElement::create('p', [
-            'class' => 'line-chart-error preformatted',
-            'data-message-nodata' => $this->translate('No data received'),
-            'data-message-error' => $this->translate('Error while fetching performance data'),
-        ]);
-
-        $config = ModuleConfig::getConfig();
-
         // Add a headline and all other elements to our element.
         $header = Html::tag('h2', $this->translate('Performance Data Graph'));
-        $header->add(new Icon('spinner', ['class' => 'spinner']));
 
         $main->add($header);
-        $main->add($error);
 
-        $charts->add((new QuickActions($config['default_timerange'])));
-        $charts->add($chart);
+        // Load the module's configuration.
+        $config = ModuleConfig::getConfigWithDefaults();
+
+        $duration = $config['default_timerange'];
+
+        // When there is a parameter for the duration we use that instead.
+        if (Url::fromRequest()->hasParam('perfdatagraphs.duration')) {
+            $duration = Url::fromRequest()->getParam('perfdatagraphs.duration');
+        }
+
+        $source = new PerfdataSource($config);
+
+        $cacheDurationInSeconds = $config['cache_lifetime'];
+        $h = $isHostCheck ? 'true': 'false';
+        // base64 since there can be whatever in the names
+        $cacheKey = base64_encode($hostName . $serviceName . $checkCommandName . $duration . $h);
+
+        // Get data from cache if it is available
+        $datasets = $source->getDataFromCache($cacheKey, $cacheDurationInSeconds);
+
+        // If not, fetch the perfdata for a given object via the hook.
+        if (!$datasets) {
+            $perfdata = $source->fetchDataViaHook($hostName, $serviceName, $checkCommandName, $duration, $isHostCheck);
+
+            // Error handling, if this gets too long, we could move this to a method.
+            if ($perfdata->isEmpty()) {
+                $msg = $this->translate('No data received');
+                $main->add(HtmlElement::create(
+                    'p',
+                    ['class' => 'line-chart-error preformatted'],
+                    $msg,
+                ));
+                return $main;
+            }
+
+            if ($perfdata->hasErrors()) {
+                $msg = $this->translate('Error while fetching performance data: %s');
+                $main->add(HtmlElement::create(
+                    'p',
+                    ['class' => 'line-chart-error preformatted'],
+                    sprintf($msg, join(' ', $perfdata->errors)),
+                ));
+                return $main;
+            }
+
+            if (!$perfdata->isValid()) {
+                $msg = $this->translate('Invalid data received: %s');
+                $main->add(HtmlElement::create(
+                    'p',
+                    ['class' => 'line-chart-error preformatted'],
+                    sprintf($msg, join(' ', $perfdata->errors)),
+                ));
+                return $main;
+            }
+
+            foreach ($perfdata->getDatasets() as $dataset) {
+                $datasets[$dataset->getTitle()] = Json::sanitize($dataset);
+            }
+
+            // After transforming the data store it. We're just storing the acutal datasets
+            // since the rest is just relevant for the request.
+            $source->storeDataToCache($cacheKey, $datasets);
+        }
+
+        $charts->add((new QuickActions(Url::fromRequest())));
+
+        // Elements in which the charts will get rendered.
+        // We use attributes on this elements to transport data
+        // to the JavaScript part of this module.
+        foreach ($datasets as $title => $data) {
+            $chart = HtmlElement::create('div', [
+                'class' => 'line-chart',
+                'id' => $elemID . '_' . $title,
+                'data-perfdata' => $data,
+            ]);
+
+            $charts->add($chart);
+        }
 
         $main->add($charts);
+
+        // We only need the toggle button when there are more charts
+        if (count($datasets) > 1) {
+            $chartsControl->add($toggleButton);
+        }
+
         $main->add($chartsControl);
 
         return $main;