+++ library/Perfdatagraphs/Icingadb/IcingaObjectHelper.php
@@ -6,6 +6,7 @@ use Icinga\Exception\NotFoundError;
 
 use Icinga\Module\Icingadb\Common\Auth;
 use Icinga\Module\Icingadb\Common\Database;
+use Icinga\Module\Icingadb\Common\Macros;
 use Icinga\Module\Icingadb\Model\Host;
 use Icinga\Module\Icingadb\Model\Service;
 
@@ -14,12 +15,13 @@ use ipl\Stdlib\Filter;
 use ipl\Orm\Model;
 
 /**
- * CustomVarsHelper is a helper class to work with custom variables of Icinga objects.
+ * IcingaObjectHelper is a helper class to work with Icinga objects.
  */
-class CustomVarsHelper
+class IcingaObjectHelper
 {
     use Database;
     use Auth;
+    use Macros;
 
     // Name of all the custom variables we use.
     public const CUSTOM_VAR_CONFIG_PREFIX  = 'perfdatagraphs_config';