+++ library/Perfdatagraphs/Model/PerfdataSet.php
@@ -22,6 +22,9 @@ class PerfdataSet implements JsonSerializable
      /** @var string The stroke of this dataset */
     protected string $stroke;
 
+     /** @var string Display this dataset or not */
+    protected bool $show = true;
+
     /** @var iterable The timstamps for this dataset */
     protected iterable $timestamps = [];
 
@@ -38,6 +41,9 @@ class PerfdataSet implements JsonSerializable
         $this->unit = $unit;
     }
 
+    /**
+     * jsonSerialize implements JsonSerializable
+     */
     public function jsonSerialize(): mixed
     {
         $d = [];
@@ -56,6 +62,10 @@ class PerfdataSet implements JsonSerializable
 
         if (isset($this->stroke)) {
             $d['stroke'] = $this->stroke;
+        }
+
+        if (isset($this->show)) {
+            $d['show'] = $this->show;
         }
 
         if (isset($this->timestamps)) {
@@ -115,27 +125,68 @@ class PerfdataSet implements JsonSerializable
         return true;
     }
 
+    /**
+     * getTitle gets the title of this dataset.
+     * @return string The title of this dataset
+     */
     public function getTitle(): string
     {
         return $this->title;
     }
 
     /**
-     * setUnit sets the unit for this data series.
-     * @param string $u
+     * setTitle sets the title of this dataset.
+     * @param string $title The title to set
      */
-    public function setUnit(string $u): void
+    public function setTitle(string $title): void
     {
-        $this->unit = $u;
+        $this->title = $title;
+    }
+
+    /**
+     * Gets the unit of this dataset.
+     *
+     * @return string The unit
+     */
+    public function getUnit(): string
+    {
+        return $this->unit;
+    }
+
+    /**
+     * setUnit sets the unit for this data series.
+     * @param string $unit The unit to set
+     */
+    public function setUnit(string $unit): void
+    {
+        $this->unit = $unit;
+    }
+
+    /**
+     * getFill gets the fill of this dataset.
+     * @return string The fill of this dataset
+     */
+    public function getFill(): string
+    {
+        return $this->fill;
     }
 
     /**
      * setFill sets the fill color of the data series.
-     * @param string $s
+     * @param string $fill
      */
-    public function setFill(string $f): void
+    public function setFill(string $fill): void
     {
-        $this->fill = $f;
+        $this->fill = $fill;
+    }
+
+    /**
+     * getStroke gets the stroke of this dataset.
+     * @return string the stroke of this dataset
+     */
+    public function getStroke(): string
+    {
+        return $this->stroke;
     }
 
     /**
@@ -148,6 +199,42 @@ class PerfdataSet implements JsonSerializable
     }
 
     /**
+     * getShow gets the show of this dataset.
+     * @return bool value of show of this dataset
+     */
+    public function getShow(): bool
+    {
+        return $this->show;
+    }
+
+    /**
+     * setShow sets the show color of the data series.
+     * @param bool $s
+     */
+    public function setShow(bool $s): void
+    {
+        $this->show = $s;
+    }
+
+    /**
+     * getSeries gets the series of this dataset.
+     * @return array The series of this dataset
+     */
+    public function getSeries(): array
+    {
+        return $this->series;
+    }
+
+    /**
+     * setSeries overrides the series of this dataset.
+     * @return array The series to set
+     */
+    public function setSeries(array $series): array
+    {
+        return $this->series = $series;
+    }
+
+    /**
      * addSeries adds a new data series to this dataset.
      * @param PerfdataSeries $s
      */
@@ -157,11 +244,29 @@ class PerfdataSet implements JsonSerializable
     }
 
     /**
+     * getTimestamps gets the timestamps of this dataset.
+     * @return iterable the timestamps of this dataset
+     */
+    public function getTimestamps(): iterable
+    {
+        return $this->timestamps;
+    }
+
+    /**
      * setTimestamps sets the timestamps for this dataset.
      * @param iterable $ts
      */
     public function setTimestamps(iterable $ts): void
     {
         $this->timestamps = $ts;
+    }
+
+    /**
+     * addTimestamp adds a timestamp to this dataset.
+     * @param mixed $timestamp The timestamp to add
+     */
+    public function addTimestamp(mixed $timestamp): void
+    {
+        $this->timestamps[] = $timestamp;
     }
 }