+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.2.0
+
+- Rework data-fetching to use PHP
+- Add human-readable units in legend
+- Extend IcingaObjectHelper to be able to resolve Macros
+- Use locale to adjust time format in legend
+- Fix inconsistent collapsible behavior during autorefresh
+- Add the option to configure custom timeranges
+
 ## v0.1.2
 
 - Fix custom variables could not be loaded in IDO